+++ cryptolyzer/tls/all.py
@@ -5,9 +5,10 @@ from collections import OrderedDict
 import attr
 
 from cryptoparser.common.algorithm import Authentication, KeyExchange
+from cryptoparser.common.utils import get_leaf_classes
 from cryptoparser.tls.version import TlsProtocolVersionFinal, TlsVersion
 
-from cryptolyzer.common.analyzer import AnalyzerTlsBase
+from cryptolyzer.common.analyzer import AnalyzerTlsBase, ProtocolHandlerTlsExactVersion
 from cryptolyzer.common.result import AnalyzerResultTls, AnalyzerTargetTls
 
 from cryptolyzer.tls.ciphers import AnalyzerCipherSuites, AnalyzerResultCipherSuites
@@ -83,6 +84,28 @@ class AnalyzerAll(AnalyzerTlsBase):
         return 'Check TLS settings of the server(s) all'
 
     @staticmethod
+    def _get_result(analyzer_class, analyzable, protocol_version):
+        for protocol_handler_class in get_leaf_classes(ProtocolHandlerTlsExactVersion):
+            if (analyzer_class in protocol_handler_class.get_analyzers() and
+                    protocol_handler_class.get_protocol_version() == protocol_version):
+                break
+        else:
+            protocol_version = None
+
+        analyzer_name = analyzer_class.get_name()
+        if protocol_version is None:
+            return {analyzer_name: None}
+
+        return {analyzer_name: analyzer_class().analyze(analyzable, protocol_version)}
+
+    @staticmethod
+    def get_versions_result(analyzable):
+        analyzer_class = AnalyzerVersions
+        analyzer_name = analyzer_class.get_name()
+
+        return {analyzer_name: analyzer_class().analyze(analyzable, None)}
+
+    @staticmethod
     def _is_key_exchange_supported(cipher_suite_results, key_exchange):
         for protocol_version, cipher_suite_result in cipher_suite_results.items():
             for cipher_suite in cipher_suite_result.cipher_suites:
@@ -100,6 +123,25 @@ class AnalyzerAll(AnalyzerTlsBase):
         return AnalyzerAll._is_key_exchange_supported(cipher_suite_results, KeyExchange.ECDHE)
 
     @staticmethod
+    def get_dhparams_result(analyzable, cipher_suite_results):
+        protocol_version = AnalyzerAll.is_dhe_supported(cipher_suite_results)
+        return AnalyzerAll._get_result(AnalyzerDHParams, analyzable, protocol_version)
+
+    @staticmethod
+    def get_curves_result(analyzable, cipher_suite_results):
+        protocol_version = AnalyzerAll.is_ecdhe_supported(cipher_suite_results)
+        return AnalyzerAll._get_result(AnalyzerCurves, analyzable, protocol_version)
+
+    @staticmethod
+    def get_pubkeyreq_result(analyzable, protocol_versions):
+        if protocol_versions:
+            protocol_version = protocol_versions[-1]
+        else:
+            protocol_version = None
+
+        return AnalyzerAll._get_result(AnalyzerPublicKeyRequest, analyzable, protocol_version)
+
+    @staticmethod
     def is_publc_key_supported(cipher_suite_results):
         for protocol_version, cipher_suite_result in cipher_suite_results.items():
             for cipher_suite in cipher_suite_result.cipher_suites:
@@ -108,42 +150,37 @@ class AnalyzerAll(AnalyzerTlsBase):
 
         return None
 
+    @staticmethod
+    def get_pubkeys_result(analyzable, cipher_suite_results):
+        protocol_version = AnalyzerAll.is_publc_key_supported(cipher_suite_results)
+        return AnalyzerAll._get_result(AnalyzerPublicKeys, analyzable, protocol_version)
+
+    @staticmethod
+    def get_sigalgos_result(analyzable, versions):
+        protocol_version = TlsProtocolVersionFinal(TlsVersion.TLS1_2)
+        if protocol_version not in versions:
+            protocol_version = None
+
+        return AnalyzerAll._get_result(AnalyzerSigAlgos, analyzable, protocol_version)
+
     def analyze(self, analyzable, protocol_version):
         results = {
             'target': AnalyzerTargetTls.from_l7_client(analyzable, protocol_version),
-            'versions': None,
         }
 
-        results['versions'] = AnalyzerVersions().analyze(analyzable, protocol_version)
-        versions = results['versions'].versions
+        results.update(self.get_versions_result(analyzable))
+        versions = results[AnalyzerVersions.get_name()].versions
 
         cipher_suite_results = OrderedDict([
             (protocol_version, AnalyzerCipherSuites().analyze(analyzable, protocol_version))
             for protocol_version in versions
         ])
 
-        if versions:
-            results['pubkeyreq'] = AnalyzerPublicKeyRequest().analyze(analyzable, versions[-1])
-        protocol_version = self.is_publc_key_supported(cipher_suite_results)
-        if protocol_version is None:
-            results['pubkeys'] = None
-        else:
-            results['pubkeys'] = AnalyzerPublicKeys().analyze(analyzable, protocol_version)
+        results.update(self.get_pubkeyreq_result(analyzable, versions))
+        results.update(self.get_pubkeys_result(analyzable, cipher_suite_results))
+        results.update(self.get_dhparams_result(analyzable, cipher_suite_results))
+        results.update(self.get_curves_result(analyzable, cipher_suite_results))
+        results.update(self.get_sigalgos_result(analyzable, versions))
+        results.update({AnalyzerCipherSuites.get_name(): list(cipher_suite_results.values())})
 
-        protocol_version = self.is_dhe_supported(cipher_suite_results)
-        if protocol_version is None:
-            results['dhparams'] = None
-        else:
-            results['dhparams'] = AnalyzerDHParams().analyze(analyzable, protocol_version)
-        protocol_version = self.is_ecdhe_supported(cipher_suite_results)
-        if protocol_version is None:
-            results['curves'] = None
-        else:
-            results['curves'] = AnalyzerCurves().analyze(analyzable, protocol_version)
-        if TlsProtocolVersionFinal(TlsVersion.TLS1_2) in versions:
-            results['sigalgos'] = AnalyzerSigAlgos().analyze(analyzable, TlsProtocolVersionFinal(TlsVersion.TLS1_2))
-        else:
-            results['sigalgos'] = None
-
-        results['ciphers'] = list(cipher_suite_results.values())
         return AnalyzerResultAll(**results)