+++ cryptolyzer/tls/analyzer.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
 
-import abc
-
 from cryptoparser.tls.version import TlsProtocolVersionFinal
 from cryptoparser.tls.version import TlsVersion
 from cryptoparser.tls.version import SslProtocolVersion
 
-from cryptolyzer.common.analyzer import ProtocolHandlerBase
+from cryptolyzer.common.analyzer import ProtocolHandlerTlsBase, ProtocolHandlerTlsExactVersion
 from cryptolyzer.tls.pubkeys import AnalyzerPublicKeys
 from cryptolyzer.tls.pubkeyreq import AnalyzerPublicKeyRequest
 from cryptolyzer.tls.ciphers import AnalyzerCipherSuites
@@ -15,28 +13,6 @@ from cryptolyzer.tls.dhparams import AnalyzerDHParams
 from cryptolyzer.tls.sigalgos import AnalyzerSigAlgos
 from cryptolyzer.tls.versions import AnalyzerVersions
 from cryptolyzer.tls.all import AnalyzerAll
-
-
-class ProtocolHandlerTlsBase(ProtocolHandlerBase):
-    @classmethod
-    @abc.abstractmethod
-    def _get_protocol_version(cls):
-        raise NotImplementedError()
-
-    @classmethod
-    def get_protocol(cls):
-        return cls._get_protocol_version().identifier
-
-    @classmethod
-    def _get_analyzer_args(cls):
-        return ([], {'protocol_version': cls._get_protocol_version()})
-
-
-class ProtocolHandlerTlsExactVersion(ProtocolHandlerTlsBase):
-    @classmethod
-    @abc.abstractmethod
-    def _get_protocol_version(cls):
-        raise NotImplementedError()
 
 
 class ProtocolHandlerSsl2(ProtocolHandlerTlsExactVersion):
@@ -48,7 +24,7 @@ class ProtocolHandlerSsl2(ProtocolHandlerTlsExactVersion):
         )
 
     @classmethod
-    def _get_protocol_version(cls):
+    def get_protocol_version(cls):
         return SslProtocolVersion()
 
 
@@ -58,7 +34,7 @@ class ProtocolHandlerSsl3(ProtocolHandlerTlsExactVersion):
         return ProtocolHandlerSsl2.get_analyzers()
 
     @classmethod
-    def _get_protocol_version(cls):
+    def get_protocol_version(cls):
         return TlsProtocolVersionFinal(TlsVersion.SSL3)
 
 
@@ -72,7 +48,7 @@ class ProtocolHandlerTls10(ProtocolHandlerTlsExactVersion):
         )
 
     @classmethod
-    def _get_protocol_version(cls):
+    def get_protocol_version(cls):
         return TlsProtocolVersionFinal(TlsVersion.TLS1_0)
 
 
@@ -82,7 +58,7 @@ class ProtocolHandlerTls11(ProtocolHandlerTlsExactVersion):
         return ProtocolHandlerTls10.get_analyzers()
 
     @classmethod
-    def _get_protocol_version(cls):
+    def get_protocol_version(cls):
         return TlsProtocolVersionFinal(TlsVersion.TLS1_1)
 
 
@@ -94,8 +70,24 @@ class ProtocolHandlerTls12(ProtocolHandlerTlsExactVersion):
         )
 
     @classmethod
-    def _get_protocol_version(cls):
+    def get_protocol_version(cls):
         return TlsProtocolVersionFinal(TlsVersion.TLS1_2)
+
+
+class ProtocolHandlerTls13(ProtocolHandlerTlsExactVersion):
+    @classmethod
+    def get_analyzers(cls):
+        analyzers = ProtocolHandlerTls12.get_analyzers()
+
+        # Temporarily remove analyzers need encrypted packets
+        return tuple(filter(
+            lambda analyzer: analyzer not in [AnalyzerPublicKeys, AnalyzerPublicKeyRequest, ],
+            analyzers
+        ))
+
+    @classmethod
+    def get_protocol_version(cls):
+        return TlsProtocolVersionFinal(TlsVersion.TLS1_3)
 
 
 class ProtocolHandlerTlsVersionIndependent(ProtocolHandlerTlsBase):
@@ -111,5 +103,5 @@ class ProtocolHandlerTlsVersionIndependent(ProtocolHandlerTlsBase):
         return 'tls'
 
     @classmethod
-    def _get_protocol_version(cls):
+    def get_protocol_version(cls):
         return None