+++ cryptolyzer/tls/pubkeyreq.py
@@ -6,7 +6,7 @@ import attr
 
 import asn1crypto.x509
 
-from cryptoparser.tls.extension import TlsSignatureAndHashAlgorithm
+from cryptoparser.tls.algorithm import TlsSignatureAndHashAlgorithm
 from cryptoparser.tls.subprotocol import (
     TlsAlertDescription,
     TlsClientCertificateType,