+++ test/tls/test_versions.py
@@ -6,7 +6,9 @@ except ImportError:
     import mock
 
 from cryptoparser.tls.subprotocol import TlsAlertDescription, SslErrorType
-from cryptoparser.tls.version import TlsVersion, TlsProtocolVersionFinal, SslProtocolVersion
+from cryptoparser.tls.version import TlsVersion, TlsProtocolVersionDraft, TlsProtocolVersionFinal, SslProtocolVersion
+
+from cryptolyzer.common.exception import SecurityError, SecurityErrorType
 
 from cryptolyzer.tls.client import L7ClientTlsBase, SslError
 from cryptolyzer.tls.exception import TlsAlert
@@ -39,6 +41,20 @@ class TestSslVersions(TestTlsCases.TestTlsBase):
             self.get_result('badssl.com', 443)
         self.assertEqual(context_manager.exception.error, SslErrorType.NO_CERTIFICATE_ERROR)
 
+    @mock.patch.object(
+        L7ClientTlsBase, 'do_ssl_handshake',
+        side_effect=SecurityError(SecurityErrorType.UNPARSABLE_MESSAGE)
+    )
+    def test_error_security_error(self, _):
+        self.assertEqual(
+            self.get_result('badssl.com', 443).versions,
+            [
+                TlsProtocolVersionFinal(TlsVersion.TLS1_0),
+                TlsProtocolVersionFinal(TlsVersion.TLS1_1),
+                TlsProtocolVersionFinal(TlsVersion.TLS1_2),
+            ]
+        )
+
     def test_ssl_2(self):
         threaded_server = self.create_server(TlsServerConfiguration(
             protocol_versions=[TlsProtocolVersionFinal(TlsVersion.SSL3), ],
@@ -51,7 +67,7 @@ class TestSslVersions(TestTlsCases.TestTlsBase):
 
     def test_versions(self):
         threaded_server = L7ServerTlsTest(
-            L7ServerTls('localhost', 0, timeout=0.2),
+            L7ServerTls('localhost', 0, timeout=0.5),
         )
         threaded_server.start()
 
@@ -71,7 +87,10 @@ class TestSslVersions(TestTlsCases.TestTlsBase):
         result = self.get_result('www.google.com', 443)
         self.assertEqual(
             result.versions,
-            [TlsProtocolVersionFinal(version) for version in [TlsVersion.TLS1_0, TlsVersion.TLS1_1, TlsVersion.TLS1_2]]
+            [
+                TlsProtocolVersionFinal(version)
+                for version in [TlsVersion.TLS1_0, TlsVersion.TLS1_1, TlsVersion.TLS1_2, TlsVersion.TLS1_3, ]
+            ]
         )
 
 
@@ -113,6 +132,29 @@ class TestTlsVersions(TestTlsCases.TestTlsBase):
         self.assertEqual(
             self.get_result('badssl.com', 443).versions,
             [TlsProtocolVersionFinal(version) for version in [TlsVersion.TLS1_0, TlsVersion.TLS1_1, TlsVersion.TLS1_2]]
+        )
+
+    def test_tls_1_3(self):
+        self.assertEqual(
+            self.get_result('www.cloudflare.com', 443).versions,
+            [
+                TlsProtocolVersionFinal(version)
+                for version in [TlsVersion.TLS1_0, TlsVersion.TLS1_1, TlsVersion.TLS1_2, TlsVersion.TLS1_3, ]
+            ]
+        )
+
+    def test_tls_1_3_draft(self):
+        self.assertEqual(
+            self.get_result('www.internet.org', 443).versions,
+            [
+                TlsProtocolVersionFinal(TlsVersion.TLS1_0),
+                TlsProtocolVersionFinal(TlsVersion.TLS1_1),
+                TlsProtocolVersionFinal(TlsVersion.TLS1_2),
+                TlsProtocolVersionDraft(23),
+                TlsProtocolVersionDraft(26),
+                TlsProtocolVersionDraft(28),
+                TlsProtocolVersionFinal(TlsVersion.TLS1_3),
+            ]
         )
 
     def test_ecdsa_only(self):